refactor: color is an enum
